Updated Asteroid Class
I made updates for the Asteroid class.

I changed the getters and setters for all List attributes:
GETTERS: return the whole list. We can use the build-in methods of Lists to access or change the elements in the list.
SETTERS: all setters accept a single object as a parameter and add it to the list.